# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/indicators": {
      "get": {
        "operationId": "AdminController_indicadores",
        "summary": "Indicadores generales (v_admin_indicadores)",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Administración"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/sales-by-city": {
      "get": {
        "operationId": "AdminController_ventas",
        "summary": "Ventas por ciudad (v_ventas_por_ciudad)",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Administración"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/top-accommodations": {
      "get": {
        "operationId": "AdminController_top",
        "summary": "Top 10 alojamientos (v_top_alojamientos)",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Administración"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/accommodations": {
      "get": {
        "operationId": "AdminController_alojamientos",
        "summary": "Todos los alojamientos (filtro por estado y búsqueda por nombre, ciudad o código)",
        "parameters": [
          {
            "name": "estado",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "BORRADOR",
                "PUBLICADO",
                "SUSPENDIDO"
              ],
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Busca en nombre, ciudad o código",
            "schema": {
              "maxLength": 80,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Administración"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/accommodations/{codigo}/status": {
      "patch": {
        "operationId": "AdminController_estadoAlojamiento",
        "summary": "Suspender o reactivar un alojamiento",
        "parameters": [
          {
            "name": "codigo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstadoAlojamientoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Administración"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/users": {
      "get": {
        "operationId": "AdminController_usuarios",
        "summary": "Usuarios",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Busca en correo, nombres o apellidos",
            "schema": {
              "maxLength": 80,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Administración"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/users/{id}/status": {
      "patch": {
        "operationId": "AdminController_estadoUsuario",
        "summary": "Activar o desactivar un usuario",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstadoUsuarioDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Administración"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/catalogs": {
      "get": {
        "operationId": "AdminController_catalogos",
        "summary": "Catálogos: tipos, amenidades, políticas, ciudades, aeropuertos",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Administración"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/catalogs/amenities": {
      "post": {
        "operationId": "AdminController_amenidad",
        "summary": "Agregar amenidad",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AmenidadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Administración"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/catalogs/types": {
      "post": {
        "operationId": "AdminController_tipo",
        "summary": "Agregar tipo de alojamiento",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TipoAlojamientoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Administración"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/catalogs/cities": {
      "post": {
        "operationId": "AdminController_ciudad",
        "summary": "Agregar ciudad",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CiudadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Administración"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/catalogs/airports": {
      "post": {
        "operationId": "AdminController_aeropuerto",
        "summary": "Agregar aeropuerto",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AeropuertoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Administración"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/taxes": {
      "get": {
        "operationId": "AdminController_impuestos",
        "summary": "Tarifas de IVA/servicio y feriados",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Administración"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "AdminController_crearImpuesto",
        "summary": "Registrar tarifa (p. ej. feriado con IVA 8% que requiere registro de turismo)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImpuestoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Administración"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/taxes/{id}": {
      "patch": {
        "operationId": "AdminController_cerrarImpuesto",
        "summary": "Cerrar la vigencia de una tarifa",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CerrarImpuestoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Administración"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/events": {
      "get": {
        "operationId": "AdminController_eventos",
        "summary": "Eventos de negocio (outbox)",
        "parameters": [
          {
            "name": "tipo",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 60,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Administración"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/jobs/{nombre}/ejecutar": {
      "post": {
        "operationId": "AdminController_ejecutar",
        "summary": "Ejecutar un job: expirar-solicitudes, completar-estancias, publicar-outbox, purgar-idempotencia",
        "parameters": [
          {
            "name": "nombre",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Administración"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Posada EC · Alojamientos API",
    "description": "Implementación del contrato \"GDS Alojamientos Core API\" (rutas del YAML bajo /api/v1) y rutas propias (auth, cuenta, anfitrión, admin). El ownerId de cada reserva se infiere del `sub` del JWT. OAuth2 simplificado: el token se obtiene con POST /api/v1/auth/login y se envía como Bearer.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Este servidor"
    }
  ],
  "components": {
    "securitySchemes": {
      "OAuth2Security": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://auth.booking-hub.com/oauth2/authorize",
            "tokenUrl": "https://auth.booking-hub.com/oauth2/token",
            "scopes": {
              "alojamientos:read": "Leer información de alojamientos, disponibilidades y reservas",
              "alojamientos:book": "Crear y alterar reservas",
              "alojamientos:cancel": "Cancelar reservas",
              "alojamientos:webhooks": "Gestionar webhooks"
            }
          },
          "clientCredentials": {
            "tokenUrl": "https://auth.booking-hub.com/oauth2/token",
            "scopes": {
              "alojamientos:read": "Leer información de alojamientos, disponibilidades y reservas",
              "alojamientos:book": "Crear y alterar reservas",
              "alojamientos:cancel": "Cancelar reservas",
              "alojamientos:webhooks": "Gestionar webhooks"
            }
          }
        },
        "description": "OAuth2 del contrato. En este prototipo el token Bearer se emite en /api/v1/auth/login."
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "EstadoAlojamientoDto": {
        "type": "object",
        "properties": {
          "estado": {
            "type": "string",
            "enum": [
              "PUBLICADO",
              "SUSPENDIDO"
            ]
          }
        },
        "required": [
          "estado"
        ]
      },
      "EstadoUsuarioDto": {
        "type": "object",
        "properties": {
          "activo": {
            "type": "boolean"
          }
        },
        "required": [
          "activo"
        ]
      },
      "AmenidadDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "minLength": 2,
            "maxLength": 60
          },
          "categoria": {
            "type": "string",
            "minLength": 2,
            "maxLength": 30
          }
        },
        "required": [
          "nombre",
          "categoria"
        ]
      },
      "TipoAlojamientoDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50
          }
        },
        "required": [
          "nombre"
        ]
      },
      "CiudadDto": {
        "type": "object",
        "properties": {
          "provincia": {
            "type": "string",
            "minLength": 3,
            "maxLength": 40
          },
          "nombre": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80
          },
          "latitud": {
            "type": "number",
            "minimum": -5,
            "maximum": 2
          },
          "longitud": {
            "type": "number",
            "minimum": -93,
            "maximum": -75
          }
        },
        "required": [
          "provincia",
          "nombre"
        ]
      },
      "AeropuertoDto": {
        "type": "object",
        "properties": {
          "codigo_iata": {
            "type": "string",
            "pattern": "/^[A-Z]{3}$/"
          },
          "nombre": {
            "type": "string",
            "minLength": 3,
            "maxLength": 120
          },
          "ciudad_id": {
            "type": "number",
            "minimum": 1
          },
          "latitud": {
            "type": "number",
            "minimum": -5,
            "maximum": 2
          },
          "longitud": {
            "type": "number",
            "minimum": -93,
            "maximum": -75
          }
        },
        "required": [
          "codigo_iata",
          "nombre",
          "ciudad_id",
          "latitud",
          "longitud"
        ]
      },
      "ImpuestoDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "minLength": 3,
            "maxLength": 80
          },
          "tipo": {
            "type": "string",
            "enum": [
              "IVA",
              "SERVICIO"
            ]
          },
          "porcentaje": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "vigente_desde": {
            "type": "string",
            "pattern": "FECHA"
          },
          "vigente_hasta": {
            "type": "string",
            "pattern": "FECHA"
          },
          "requiere_registro_turismo": {
            "type": "boolean"
          },
          "estrellas_minimas": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          }
        },
        "required": [
          "nombre",
          "tipo",
          "porcentaje",
          "vigente_desde"
        ]
      },
      "CerrarImpuestoDto": {
        "type": "object",
        "properties": {
          "vigente_hasta": {
            "type": "string",
            "pattern": "FECHA"
          }
        },
        "required": [
          "vigente_hasta"
        ]
      }
    }
  }
}